incusd/scriptlet: refactor marshal
