+++ internal/server/scriptlet/marshal/marshal_test.go
@@ -1,4 +1,4 @@
-package scriptlet
+package marshal
 
 import (
 	"fmt"
